+++ ft_putnbr.c
@@ -6,16 +6,16 @@
 /*   By: namoreir <namoreir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/23 10:30:22 by namoreir          #+#    #+#             */
-/*   Updated: 2023/08/23 16:11:16 by namoreir         ###   ########.fr       */
+/*   Updated: 2023/08/24 12:37:46 by namoreir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_putnbr(int nb)
+int	ft_putnbr(int nb)
 {
 	long int	i;
-	size_t size;
+	int			size;
 
 	i = nb;
 	size = 0;
@@ -31,9 +31,9 @@ size_t	ft_putnbr(int nb)
 	return (size);
 }
 
-size_t	ft_putnbr_uns(unsigned int nb)
+int	ft_putnbr_uns(unsigned int nb)
 {
-	size_t size;
+	int	size;
 
 	size = 0;
 	if (nb >= 10)
@@ -53,5 +53,4 @@ size_t	ft_putnbr_uns(unsigned int nb)
 // 	int result_2 = printf("%d", i);
 // 	printf("%c", '\n');
 // 	printf("%d", result_2);
-	
 // }
